moves storageAgents and mockStorageAgents to agents folder, removes unused imports from query handlers
