+++ examples/bri-3/src/bri/identity/bpiSubjects/api/subjects.controller.spec.ts
@@ -7,8 +7,8 @@ import { DeleteBpiSubjectCommandHandler } from '../capabilities/deleteBpiSubject
 import { GetAllBpiSubjectsQueryHandler } from '../capabilities/getAllBpiSubjects/getAllBpiSubjectsQuery.handler';
 import { GetBpiSubjectByIdQueryHandler } from '../capabilities/getBpiSubjectById/getBpiSubjectByIdQuery.handler';
 import { UpdateBpiSubjectCommandHandler } from '../capabilities/updateBpiSubject/updateBpiSubjectCommand.handler';
-import { BpiSubjectStorageAgent } from '../persistence/bpiSubjectsStorage.agent';
-import { MockBpiSubjectStorageAgent } from '../persistence/mockBpiSubjectStorage.agent';
+import { BpiSubjectStorageAgent } from '../agents/bpiSubjectsStorage.agent';
+import { MockBpiSubjectStorageAgent } from '../agents/mockBpiSubjectStorage.agent';
 import { CreateBpiSubjectDto } from './dtos/request/createBpiSubject.dto';
 import { UpdateBpiSubjectDto } from './dtos/request/updateBpiSubject.dto';
 import { NAME_EMPTY_ERR_MESSAGE, NOT_FOUND_ERR_MESSAGE } from './err.messages';
