+++ examples/bri-3/src/bri/identity/bpiSubjects/capabilities/getBpiSubjectById/getBpiSubjectByIdQuery.handler.ts
@@ -1,8 +1,6 @@
-import { NotFoundException } from "@nestjs/common";
 import { IQueryHandler, QueryHandler } from "@nestjs/cqrs";
 import { BpiSubjectDto } from "../../api/dtos/response/bpiSubject.dto";
-import { NOT_FOUND_ERR_MESSAGE } from "../../api/err.messages";
-import { BpiSubjectStorageAgent } from "../../persistence/bpiSubjectsStorage.agent";
+import { BpiSubjectStorageAgent } from "../../agents/bpiSubjectsStorage.agent";
 import { GetBpiSubjectByIdQuery } from "./getBpiSubjectById.query";
 
 @QueryHandler(GetBpiSubjectByIdQuery)
@@ -13,10 +11,6 @@ export class GetBpiSubjectByIdQueryHandler implements IQueryHandler<GetBpiSubjec
 
   async execute(query: GetBpiSubjectByIdQuery) {
     const bpiSubject = await this.storageAgent.getBpiSubjectById(query.id);
-
-    if (!bpiSubject) {
-      throw new NotFoundException(NOT_FOUND_ERR_MESSAGE)
-    }
 
     return { // TODO: Write generic mapper domainObject -> DTO
         id: bpiSubject.id,
