+++ examples/bri-3/src/bri/workgroup/worksteps/capabilities/getAllWorksteps/getAllWorkstepsQuery.handler.ts
@@ -1,6 +1,6 @@
 import { IQueryHandler, QueryHandler } from "@nestjs/cqrs";
 import { WorkstepDto } from "../../api/dtos/response/workstep.dto";
-import { WorkstepStorageAgent } from "../../persistence/workstepsStorage.agent";
+import { WorkstepStorageAgent } from "../../agents/workstepsStorage.agent";
 import { GetAllWorkstepsQuery } from "./getAllWorksteps.query";
 
 @QueryHandler(GetAllWorkstepsQuery)
